chore: move misc plugins to test/csrc/plugin
