+++ src/test/csrc/plugin/tllogger/tllogger.cpp
@@ -1,4 +1,4 @@
-#include "logger.h"
+#include "tllogger.h"
 
 bool dump_tl;
 sqlite3 *mem_db;
